Move CreateKeyAsync, GetKeyMetadataAsync and SetKeyMetadataAsync from Signus to Crypto and update tests.
