+++ wrappers/dotnet/indy-sdk-dotnet-test/CryptoTests/CreateKeyTest.cs
@@ -1,9 +1,9 @@
-﻿using Hyperledger.Indy.SignusApi;
+﻿using Hyperledger.Indy.CryptoApi;
 using Hyperledger.Indy.Test.Util.Base58Check;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Threading.Tasks;
 
-namespace Hyperledger.Indy.Test.SignusTests
+namespace Hyperledger.Indy.Test.CryptoTests
 {
     [TestClass]
     public class CreateKeyTest : IndyIntegrationTestWithSingleWallet
@@ -12,14 +12,14 @@ namespace Hyperledger.Indy.Test.SignusTests
         public async Task TestCreateKeyWorksForSeed()
         {
             var paramJson = string.Format("{{\"seed\":\"{0}\"}}", MY1_SEED);
-            var senderVk = await Signus.CreateKeyAsync(wallet, paramJson);
+            var senderVk = await Crypto.CreateKeyAsync(wallet, paramJson);
             Assert.AreEqual(32, Base58CheckEncoding.DecodePlain(senderVk).Length);
         }
 
         [TestMethod]
         public async Task TestCreateKeyWorksWithoutSeed()
         {
-            var senderVk = await Signus.CreateKeyAsync(wallet, "{}");
+            var senderVk = await Crypto.CreateKeyAsync(wallet, "{}");
             Assert.AreEqual(32, Base58CheckEncoding.DecodePlain(senderVk).Length);
         }
 
@@ -29,7 +29,7 @@ namespace Hyperledger.Indy.Test.SignusTests
             var paramJson = string.Format("{{\"seed\":\"{0}\"}}", "invalidSeedLength");
 
             var ex = await Assert.ThrowsExceptionAsync<InvalidStructureException>(() =>
-               Signus.CreateKeyAsync(wallet, paramJson)
+               Crypto.CreateKeyAsync(wallet, paramJson)
            );
         }
     }
