+++ wrappers/dotnet/indy-sdk-dotnet-test/CryptoTests/GetKeyMetadataTest.cs
@@ -1,12 +1,9 @@
-﻿using Hyperledger.Indy.SignusApi;
+﻿using Hyperledger.Indy.CryptoApi;
 using Hyperledger.Indy.WalletApi;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace Hyperledger.Indy.Test.SignusTests
+namespace Hyperledger.Indy.Test.CryptoTests
 {
     [TestClass]
     public class GetKeyMetadataTest : IndyIntegrationTestWithSingleWallet
@@ -14,16 +11,16 @@ namespace Hyperledger.Indy.Test.SignusTests
         [TestMethod]
         public async Task TestGetKeyMetadataWorks()
         {
-            await Signus.SetKeyMetadataAsync(wallet, VERKEY, METADATA);
-            var receivedMetadata = await Signus.GetKeyMetadataAsync(wallet, VERKEY);
+            await Crypto.SetKeyMetadataAsync(wallet, VERKEY, METADATA);
+            var receivedMetadata = await Crypto.GetKeyMetadataAsync(wallet, VERKEY);
             Assert.AreEqual(METADATA, receivedMetadata);
         }
 
         [TestMethod]
         public async Task TestGetKeyMetadataWorksForEmptyString()
         {
-            await Signus.SetKeyMetadataAsync(wallet, VERKEY, string.Empty);
-            var receivedMetadata = await Signus.GetKeyMetadataAsync(wallet, VERKEY);
+            await Crypto.SetKeyMetadataAsync(wallet, VERKEY, string.Empty);
+            var receivedMetadata = await Crypto.GetKeyMetadataAsync(wallet, VERKEY);
             Assert.AreEqual(string.Empty, receivedMetadata);
         }
 
@@ -31,7 +28,7 @@ namespace Hyperledger.Indy.Test.SignusTests
         public async Task TestGetKeyMetadataWorksForNoMetadata()
         {
             var ex = await Assert.ThrowsExceptionAsync<WalletValueNotFoundException>(() =>
-               Signus.GetKeyMetadataAsync(wallet, VERKEY)
+               Crypto.GetKeyMetadataAsync(wallet, VERKEY)
            );
         }
     }
